+++ packages/codemirror/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.4"></a>
+## [2.0.4](https://github.com/hpcc-systems/Visualization/compare/@hpcc-js/codemirror@2.0.3...@hpcc-js/codemirror@2.0.4) (2018-09-25)
+
+**Note:** Version bump only for package @hpcc-js/codemirror
+
+
+
+
+
 <a name="2.0.3"></a>
 ## [2.0.3](https://github.com/hpcc-systems/Visualization/compare/@hpcc-js/codemirror@2.0.2...@hpcc-js/codemirror@2.0.3) (2018-09-24)